refactor(codegen): extract std::env::* into stdlib/env.rs (Round 1.2)
5 fns lifted into EnvStdlib<'ctx> trait: args_count, arg, arg_or,
var, var_exists. No new pub(crate) surface needed beyond Round 1.1.

codegen.rs: 44690 -> 44450 (-240 lines).
